+++ src/main/java/ru/practicum/shareit/booking/AvailabilityStatus.java
@@ -1,0 +1,15 @@
+package ru.practicum.shareit.booking;
+
+/**
+ * ENUM of booking item status
+ * WAITING - new booking, waiting for approving
+ * APPROVED - booking approved by owner
+ * REJECTED - booking has rejected by owner
+ * CANCELED - booking canceled by creator of response
+ */
+public enum AvailabilityStatus {
+    WAITING,
+    APPROVED,
+    REJECTED,
+    CANCELED
+}
