+++ src/main/java/ru/practicum/shareit/booking/dto/BookingDto.java
@@ -3,8 +3,7 @@ package ru.practicum.shareit.booking.dto;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
-import ru.practicum.shareit.booking.Booking;
-import ru.practicum.shareit.booking.Status;
+import ru.practicum.shareit.booking.AvailabilityStatus;
 
 import java.time.LocalDateTime;
 
@@ -19,15 +18,6 @@ public class BookingDto {
     private LocalDateTime end;
     private String item;
     private String owner;
-    private Status status;
+    private AvailabilityStatus status;
 
-    public BookingDto toDto(Booking booking) {
-        return BookingDto.builder()
-                .start(booking.getStart())
-                .end(booking.getEnd())
-                .item(booking.getItem().getName())
-                .owner(booking.getOwner().getName())
-                .status(booking.getStatus())
-                .build();
-    }
 }
