+++ src/main/java/ru/practicum/shareit/request/RequestMapper.java
@@ -1,0 +1,14 @@
+package ru.practicum.shareit.request;
+
+import ru.practicum.shareit.request.dto.ItemRequestDto;
+
+public class RequestMapper {
+
+    public ItemRequestDto toDto(ItemRequest itemRequest) {
+        return ItemRequestDto.builder()
+                .description(itemRequest.getDescription())
+                .requesterName(itemRequest.getRequestor().getName())
+                .created(itemRequest.getCreated())
+                .build();
+    }
+}
